Create a custom injected class for claims injection
